configure.ac: Use AC_CHECK_FUNCS([malloc]) instead of AC_FUNC_MALLOC
AC_FUNC_MALLOC will define malloc to rpl_malloc in case libc is not the
Gnu one. Since most probable replacements (uClibc, musl) are behaving
the same regarding the malloc(0) case and glslsandbox-player is not really
dependant on this case, we can safely just check the presence of the malloc()
function.

Signed-off-by: Julien Olivain <juju@cotds.org>