native_gfx: mali: add support for Arm Mali framebuffer native windowing
This is targeting Mali GPUs present in AllWinner SoCs. Those are present in
low cost OrangePi boards for example. The code was tested on OrangePi Zero,
OrangePi Zero Plus 2 H3 and OrangePi One.

This code is meant to run with Arm Mali Framebuffer driver from:
https://github.com/linux-sunxi/sunxi-mali

The driver needs to be installed with EGL_TYPE=framebuffer

For X11 native windowing, glslsandbox-player can be compiled with
standard Mesa libraries and headers, then executed with Mali libraries.

Signed-off-by: Julien Olivain <juju@cotds.org>