build: use temporary files
Shader files are marked as .SECONDARY in Makefile.  This allow to keep
those intermediate files, preventing to download them again at the next
build.  If download or processing fails, an incomplete file is left.
To prevent this situation, we process files in temporary ones, and move
them to the actual target file, only if command succeed.

Signed-off-by: Julien Olivain <juju@cotds.org>