configure.ac: check if eglGetPlatformDisplay() is really available
Normally, eglGetPlatformDisplay() function is guaranteed to be
available if EGL version is 1.5 or greater. The call to this function
was protected by checking if EGL_VERSION_1_5 was defined.

Some GPU vendors are distributing their drivers as binary only. The
corresponding headers should be aligned to the same version of the
specification implemented by this binary driver.

In some real-world situations, those headers are installed from a
recent Mesa3D, then the binary from the GPU vendor package. If the
Mesa3D package advertise a more recent version than the implemented
version of the GPU vendor binary, the build will fail.

This case happened in Buildroot, for eglGetPlatformDisplay(), where
egl.h from Mesa3D advertise EGL 1.5 by defining EGL_VERSION_1_5, but
the libEGL.so actually implements 1.4.

The workaround is to actually check for the eglGetPlatformDisplay()
function at build configuration time.

Fixes:
http://autobuild.buildroot.net/results/a950e90d5f8405534566df5c7a8875c293cf8845/

Signed-off-by: Julien Olivain <juju@cotds.org>