native_gfx: x11: add the ability to set wm state
This option allow the user to set the window always on top,
always below, and control fullscreen.

The window stacking order can be controlled with the
GSP_X11_WIN_STACKING environment variable. If set to "above"
the window will be always on top other windows (with default
stacking order). If set to "below", the window will be below
other windows (this can be useful to put the the program
behind other transparent windows).

Fullscreen can be enabled by setting GSP_X11_FULLSCREEN
environment variable to '1'.

Closes #3

Signed-off-by: Julien Olivain <juju@cotds.org>