player: add specific support for Emscripten main loop
Emscripten needs specific care regarding the main rendering loop. Since
the code is executed from the javascript interpreter in a browser, the
normal infinite render loop would overload this interpreter. In
Emscripten, the workaround is to register a callback of a single frame
rendering instead. The browser will trigger this callback when needed
(ideally once per vsync for a smooth rendering). Since this Emscripten
loop never returns, the cleanup code and shutdown needs to be called
from this frame callback.

Signed-off-by: Julien Olivain <ju.o@free.fr>