native_gfx: wgl: move display openning code in window creation
Device context needs a reference to a window. So the first thing to do
is to actually create a window. Since we don't know the window dimension
here, we move all this code into the window creation. This fixes the
window size which was not following the dimensions passed to -W/-H
options.

Signed-off-by: Julien Olivain <ju.o@free.fr>