configure.ac: workaround old gl2.h not defining prototypes
There is some old (2016-04-28) Khronos GLES2 header files that protect
core GLES 2 function prototypes with the GL_GLEXT_PROTOTYPES macro. If
not defined, no prototype are present. That will likely result to a
compilation failure.

See the last gl2.h revision with the issue:
https://cvs.khronos.org/svn/repos/registry/trunk/public/gles/api/GLES2/gl2.h?p=33155

This issue was fixed in gl2.h svn r33156, in which the even older behavior
(of having function prototypes defined by default) was restored. This
fix in Khronos header protects core GL ES 2 function with the
GL_GLES_PROTOTYPES macro. If not defined, it'll be enabled by default.
Core GL ES 2 functions can be disabled by definined GL_GLES_PROTOTYPES
to 0.

Corrected gl2.h header:
https://cvs.khronos.org/svn/repos/registry/trunk/public/gles/api/GLES2/gl2.h?p=33156

Some version of those problematic GLES 2 headers are still in the wild.
glslsandbox-player always need those GL ES 2 function prototypes. It's
also assumed the program will be compiled with up-to-date Khronos
headers. In order to still be able to compile with those old headers
without polluting too much the code, the workaround is to detect those
old headers if the macro GL_GLES_PROTOTYPES is undefined at configure
time, and define GL_GLEXT_PROTOTYPES (with "GLEXT" and not "GLES") in
that case.

Signed-off-by: Julien Olivain <ju.o@free.fr>