native_gfx: kms: search for drm devices if possible
Enumerate DRM devices using drmGetDevices2() libdrm function, if
available. This allow to detect a suitable device without knowing
its kernel drm driver name in advance.

This drmGetDevices2() function was introduced libdrm in commit
https://gitlab.freedesktop.org/mesa/drm/-/commit/11687bf4180f7e21045ed9c4730533c40fe01ea5
This commit is included in tag libdrm-2.4.75 (2017-01-27)

When the function drmGetDevices2() is not present, the old behavior is
preserved (looping through a built-in list of known drm driver names).
In all cases, the behavior of environement variable GSP_DRM_DRIVER is
also preserved.

Signed-off-by: Julien Olivain <ju.o@free.fr>