configure.ac: fix eglGetPlatformDisplay detection
The function detection was broken. The function was detected using
the pkg-config PKG_CHECK_MODULE() autotools m4 macro. This macro does
not add the libraries in the search list (it defines libxyz_LIBS
instead). The AC_CHECK_FUNCS() macro was never finding the function.
We also don't want to call AC_CHECK_LIBS() which would globally add
the searched lib, without honoring libs given by pkg-config. As a
trade-off/workaround, we temporarily add the libs we got from
pkg-config in LIBS, and restore the original LIBS list after the
check.

Signed-off-by: Julien Olivain <ju.o@free.fr>