player: check if native windowing received an exit event in render loop
By default, the application runs an infinite render loop. In that case,
the application cleanup code would never be executed.

It is possible to limit the rendering loop to a number of frames (with
-f option) or to a runtime duration (with -t option). Before this
commit, it was the only way to let the application run its cleanup code.

This commit adds a new way to stop the rendering loop: a native
windowing system can signal the application it would like to terminate.
This is implemented with the native_gfx_request_exit() function (it
returns non-zero to request exit).

Simple native windowing systems that cannot request a program
termination implement that with a dummy function always returning zero.

Signed-off-by: Julien Olivain <ju.o@free.fr>