player: add the option to select a random shader
Before this patch, shader selection was left to the caller. If a random
shader was needed, the caller script could pick a random one from the
list. In the Android port, there is no possibility to easily have a
caller script. This patch add a special shader name "__RANDOM__" that
will randomly pick a shader from the builtin list.

Signed-off-by: Julien Olivain <ju.o@free.fr>