shader.list: use updated shaders with highp precision
Some shaders are using a trick creating an infinite loop with a for(),
and adding a break on some condition. The default precision was set to
mediump. If the break condition rely on high precision values, and the
OpenGL ES implementation has the minimum precision required by the
mediump, this can create an infinite loop in the shader core (possibly
creating GPU hangs).

This commit points to newer revision of the shaders, using a highp
precision to solve the issue.

See issue #6.

Signed-off-by: Julien Olivain <ju.o@free.fr>